requirement.txt
qiniu
clipboard
pillow